Revert "WT-13442 Do not allow out of order truncate commit timestamp (#11012)"
This reverts commit 2a2b22d3dd7e6734ecf6794f4dbd50ffcb59f26a.
