+++ src/btree/bt_vrfy.c
@@ -204,7 +204,7 @@ __wt_verify(WT_SESSION_IMPL *session, const char *cfg[])
     WT_DECL_RET;
     WT_VSTUFF *vs, _vstuff;
     size_t root_addr_size;
-    uint8_t root_addr[WT_ADDR_MAX_COOKIE];
+    uint8_t root_addr[WT_BTREE_MAX_ADDR_COOKIE];
     const char *name;
     bool bm_start, quit, skip_hs;
