+++ test/unittest/tests/block/test_block_addr.cpp
@@ -89,7 +89,7 @@ static void
 test_pack_and_unpack_addr_cookie(
   WT_BLOCK *block, wt_off_t pack_offset, uint32_t pack_size, uint32_t pack_checksum)
 {
-    uint8_t p[WT_ADDR_MAX_COOKIE], *pp;
+    uint8_t p[WT_BTREE_MAX_ADDR_COOKIE], *pp;
     pp = p;
 
     size_t addr_size;
